MarketBasketAnalysis/requirements.txt
asgiref==3.6.0
contourpy==1.0.7
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.5
django-cors-headers==3.13.0
django-import-export==3.0.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
fonttools==4.38.0
joblib==1.2.0
kiwisolver==1.4.4
MarkupPy==1.14
matplotlib==3.6.3
mlxtend==0.21.0
mysqlclient==2.1.1
numpy==1.24.1
odfpy==1.4.1
openpyxl==3.0.10
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
scikit-learn==1.2.0
scipy==1.10.0
seaborn==0.12.2
six==1.16.0
sqlparse==0.4.3
tablib==3.3.0
threadpoolctl==3.1.0
xlrd==2.0.1
xlwt==1.3.0
